requirements.txt
django==4.2
django-log-request-id==2.1.0
djangorestframework==3.14
psycopg2-binary==2.9.6
djangorestframework_simplejwt==5.2.2
# pymemcache==4.0.0
django-cors-headers==3.14.0
uwsgi==2.0.21
requests==2.30.0
python-dateutil==2.8.2